+++ subpackages/utils/source/odood/utils/zip.d
@@ -1,15 +1,13 @@
 module odood.utils.zip;
 
-private import std.logger: warningf;
 private import std.format: format;
-private import std.algorithm.searching: endsWith, startsWith;
 private import std.exception: enforce;
 
 private import thepath: Path;
 
 private import odood.exception: OdoodException;
 
-private import zipper;
+private import zipper: Zipper;
 
 /** Extract zip archive to destination directory